@@ -3,25 +3,35 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.8.5",
-    "@mui/styles": "^5.8.4",
-    "@testing-library/jest-dom": "^5.16.4",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.11.0",
+    "@mui/lab": "^5.0.0-alpha.118",
+    "@mui/material": "^5.11.7",
+    "@mui/styles": "^5.11.7",
+    "@tanstack/react-query": "^4.24.4",
+    "@tanstack/react-query-devtools": "^4.24.4",
+    "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^11.2.7",
     "@testing-library/user-event": "^12.8.3",
+    "aws-amplify": "^4.3.45",
     "axios": "^0.21.4",
-    "bootstrap": "^5.1.3",
+    "bootstrap": "^5.2.3",
     "dotenv": "^10.0.0",
-    "react": "^17.0.2",
-    "react-bootstrap": "^2.4.0",
-    "react-dom": "^17.0.2",
-    "react-router": "^5.3.3",
-    "react-router-dom": "^5.3.3",
-    "react-scripts": "4.0.3",
-    "styled-components": "^5.3.5",
-    "web-vitals": "^1.1.2"
+    "immutability-helper": "^3.1.1",
+    "js-cookie": "^3.0.1",
+    "lodash": "^4.17.21",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.0",
+    "react-dnd": "^16.0.1",
+    "react-dnd-html5-backend": "^16.0.1",
+    "react-dom": "^18.2.0",
+    "react-router": "^5.3.4",
+    "react-router-dom": "^5.3.4",
+    "react-scripts": "^5.0.1",
+    "styled-components": "^5.3.6",
+    "web-vitals": "^1.1.2",
+    "yup": "^0.32.11"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -30,7 +40,8 @@
     "eject": "react-scripts eject",
     "deploy": "aws s3 sync --profile TMG_michael --acl public-read build/ s3://southwestwatches.com",
     "deployCodeBuild": "aws s3 sync --acl public-read build/ s3://southwestwatches.com",
-    "deployCodeBuildWWW": "aws s3 sync --acl public-read build/ s3://www.southwestwatches.com"
+    "deployCodeBuildWWW": "aws s3 sync --acl public-read build/ s3://www.southwestwatches.com",
+    "lint": "eslint .eslintrc.js"
   },
   "eslintConfig": {
     "extends": [
@@ -48,5 +59,9 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "eslint": "^8.33.0",
+    "eslint-plugin-react": "^7.32.2"
   }
 }
