@@ -1,12 +1,12 @@
 import {
-  WatchSlogan,
+  Image,
   SloganBox,
   WatchDualPictureContainer,
-  Image,
+  WatchSlogan,
 } from "./style.js";
 import { CenteredParagraph } from "../../../Common/Style/commonStyle";
-
 import { Paper } from "@mui/material/";
+import React from "react";
 const Pictures = [
   {
     src: "Omega-Speedmaster-3861-2021-Unboxing-WatchGecko10.jpeg",
@@ -40,8 +40,8 @@ export default function WatchContent() {
       </CenteredParagraph>
 
       <WatchDualPictureContainer>
-        {Pictures.map((picture) => (
-          <Paper elevation={18}>
+        {Pictures.map((picture, i) => (
+          <Paper elevation={18} key={i}>
             <Image
               src={`https://southwest-watches-media.s3.amazonaws.com/${picture.src}`}
               alt={picture.alt}
