@@ -1,6 +1,7 @@
 import { BrandList, BrandsSection } from "./style.js";
-import { Paper, Card, CardContent, Typography, Grid } from "@mui/material/";
+import { Card, CardContent, Grid, Paper, Typography } from "@mui/material/";
 import { Brands } from "./brandsList";
+import React from "react";
 export default function WatchBrands() {
   return (
     <Paper elevation={8}>
@@ -14,8 +15,8 @@ export default function WatchBrands() {
             justifyContent="center"
             alignItems="center"
           >
-            {Brands.map((brand) => (
-              <Grid item xs={6} md={4}>
+            {Brands.map((brand, i) => (
+              <Grid item xs={6} md={4} key={i}>
                 <Card sx={{ height: "100%" }}>
                   <CardContent>
                     <Typography variant="h7">{brand.name}</Typography>
