@@ -1,9 +1,11 @@
+import ContactBanner from "../Common/ContactBanner/index";
 import ImageCarousel from "../Common/Carousel/index";
+import React from "react";
+import WatchBrands from "./components/WatchBrands/index";
 import { WatchCarouselImages } from "../Common/Carousel/data";
 import WatchContent from "./components/Content/index";
-import ContactBanner from "../Common/ContactBanner/index";
 import { WatchPageContainer } from "./style";
-import WatchBrands from "./components/WatchBrands/index";
+
 export default function About() {
   return (
     <WatchPageContainer>
