@@ -1,11 +1,14 @@
-import React, { useState, useEffect } from "react";
-
-import axios from "axios";
-import { useHistory } from "react-router";
-import UploadImage from "./components/UploadImage/index";
-import { makeStyles } from "@mui/styles";
-import { TextField, Paper, Button } from "@mui/material";
 import { ContactFormWrapper, WishListCard } from "./style.js";
+import { Paper, TextField } from "@mui/material";
+import { useEffect, useState } from "react";
+import { LoadingButton } from "@mui/lab";
+import React from "react";
+import UploadImage from "../../../Common/UploadImage/index";
+import { makeStyles } from "@mui/styles";
+import { sendMessageSchema } from "../../../utils/validate";
+import { useHistory } from "react-router";
+import { useSendLeadMessage } from "../../../api/index";
+const uploadImageGridSizes = { xs: 12, md: 6 };
 
 const useStyles = makeStyles(() => ({
   root: {
@@ -18,65 +21,85 @@ const useStyles = makeStyles(() => ({
   },
 }));
 
+const defaultWishListPayload = {
+  first_name: "",
+  last_name: "",
+  email: "",
+  phone: "",
+  type: "",
+  make: "",
+  model: "",
+  description: "",
+  image_URL: "",
+};
 export default function WishList() {
-  const [wishlistEntry, setWishlistEnty] = useState({});
-  const [loading, setLoading] = useState(false);
-  const [formValid, setFormValid] = useState(false);
+  const [wishlistEntry, setWishlistEntry] = useState(defaultWishListPayload);
   const [formError, setFormError] = useState(false);
-
+  const [validationErrors, setValidationErrors] = useState({});
+  const [wishlistIsEmpty, setWishlistIsEmpty] = useState(true);
   const history = useHistory();
   const classes = useStyles();
 
-  const handleSubmit = (e) => {
-    addWishlistEntry();
-  };
-
-  const addWishlistEntry = () => {
-    setLoading(true);
-    setFormError(false);
-    axios
-      .post(
-        "https://8zrqystn2h.execute-api.us-east-1.amazonaws.com/prod/wishlist",
-        {
-          first_name: wishlistEntry.first_name,
-          last_name: wishlistEntry.last_name,
-          email: wishlistEntry.email,
-          phone: wishlistEntry.phone,
-          type: wishlistEntry.type,
-          make: wishlistEntry.make,
-          model: wishlistEntry.model,
-          description: wishlistEntry.description,
-          image_URL: wishlistEntry.image_URL,
-        },
-        {
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      )
-      .then((res) => {
-        console.log(res);
-        setLoading(false);
-        history.push("/contact/success");
-        console.log("should have redirected");
-      })
-      .catch((e) => {
-        console.log(e);
-        setLoading(false);
-        setFormError(true);
-      });
-  };
+  const { mutate: sendLeadMessage, isLoading } = useSendLeadMessage({
+    onSuccess: async () => {
+      setWishlistEntry(defaultWishListPayload);
+      setValidationErrors({});
+      history.push("/contact/success");
+    },
+    onError: () => setFormError(true),
+  });
 
   useEffect(() => {
-    const isPhoneSet = !!wishlistEntry.phone;
-    const isFirstNameSet = !!wishlistEntry.first_name;
-    const isDescriptionSet = !!wishlistEntry.description;
-    const isEmailSet = !!wishlistEntry.email;
-
-    setFormValid(
-      isPhoneSet && isFirstNameSet && isDescriptionSet && isEmailSet
-    );
+    const nonEmptyValueCount = Object.values(wishlistEntry).filter(
+      (value) => value !== ""
+    ).length;
+    if (nonEmptyValueCount > 0) {
+      setWishlistIsEmpty(false);
+    } else if (nonEmptyValueCount === 0) {
+      setWishlistIsEmpty(true);
+    }
   }, [wishlistEntry]);
+
+  const isValid = async () => {
+    try {
+      await sendMessageSchema.validate(wishlistEntry, {
+        abortEarly: false,
+      });
+
+      setValidationErrors({});
+      return true;
+    } catch (e) {
+      if (e.inner) {
+        const errors = e.inner.reduce((obj, error) => {
+          obj[error.path] = error.message;
+          return obj;
+        }, {});
+        setValidationErrors(errors);
+      }
+      return false;
+    }
+  };
+
+  const handleSubmit = async () => {
+    if (await isValid()) {
+      setFormError(false);
+      sendLeadMessage({ body: wishlistEntry });
+    }
+  };
+
+  const handleChange = (e) => {
+    setWishlistEntry({
+      ...wishlistEntry,
+      [e.target.name]: e.target.value,
+    });
+  };
+
+  const handleImageChange = (url) => {
+    setWishlistEntry({
+      ...wishlistEntry,
+      image_URL: url,
+    });
+  };
 
   return (
     <ContactFormWrapper>
@@ -92,94 +115,71 @@ export default function WishList() {
           required
           variant="filled"
           color="primary"
-          type="text"
+          label="First Name"
+          name="first_name"
           value={wishlistEntry.first_name}
-          placeholder="Enter First Name"
-          onChange={(e) =>
-            setWishlistEnty({
-              ...wishlistEntry,
-              first_name: e.target.value,
-            })
-          }
-          helperText="First Name"
+          error={!!validationErrors.first_name}
+          helperText={validationErrors.first_name}
+          onChange={handleChange}
         ></TextField>
 
         <TextField
           required
           variant="filled"
           color="primary"
-          type="text"
-          placeholder="Enter Last Name"
+          name="last_name"
+          label="Last Name"
           value={wishlistEntry.last_name}
-          onChange={(e) =>
-            setWishlistEnty({
-              ...wishlistEntry,
-              last_name: e.target.value,
-            })
-          }
-          helperText="Last Name"
+          error={!!validationErrors.last_name}
+          helperText={validationErrors.last_name}
+          onChange={handleChange}
         ></TextField>
         <TextField
           required
           variant="filled"
           color="primary"
-          type="phone"
-          placeholder="Enter Phone Number"
+          label="Phone Number"
+          name="phone"
           value={wishlistEntry.phone}
-          onChange={(e) =>
-            setWishlistEnty({
-              ...wishlistEntry,
-              phone: e.target.value,
-            })
-          }
-          helperText="Phone Number"
+          error={!!validationErrors.phone}
+          helperText={validationErrors.phone}
+          onChange={handleChange}
         ></TextField>
         <TextField
           required
           variant="filled"
           color="primary"
-          type="email"
-          placeholder="Enter Email"
+          label="Email"
+          name="email"
           value={wishlistEntry.email}
-          onChange={(e) =>
-            setWishlistEnty({
-              ...wishlistEntry,
-              email: e.target.value,
-            })
-          }
-          helperText="Email"
+          error={!!validationErrors.email}
+          helperText={validationErrors.email}
+          onChange={handleChange}
         ></TextField>
         <TextField
-          required
           variant="filled"
           color="primary"
-          type="text"
-          placeholder="Enter Description"
+          name="description"
+          label="Description"
           value={wishlistEntry.description}
-          onChange={(e) =>
-            setWishlistEnty({
-              ...wishlistEntry,
-              description: e.target.value,
-            })
-          }
-          helperText="What are you looking for? Add any extra details about item. Bracelets, Materials, Dials, Bezel,
-          etc..."
+          error={!!validationErrors.description}
+          helperText={validationErrors.description}
+          onChange={handleChange}
         />
         <UploadImage
-          setWishlistEnty={setWishlistEnty}
-          {...wishlistEntry}
-          loading={loading}
-          wishlistEntry={wishlistEntry}
+          handleImageChange={handleImageChange}
+          pathKey="contactformupload"
+          gridSizes={uploadImageGridSizes}
         />
-        <Button
+        <LoadingButton
           variant="contained"
+          loading={isLoading}
           color={formError ? "error" : "primary"}
-          type="button"
-          disabled={!formValid || loading}
-          onClick={(e) => handleSubmit(e)}
+          disabled={wishlistIsEmpty}
+          onClick={() => handleSubmit()}
         >
           {formError ? "There was an error. Please try again" : "Submit"}
-        </Button>
+        </LoadingButton>
       </Paper>
     </ContactFormWrapper>
   );
