@@ -1,6 +1,6 @@
+import { ContactUsContent as Content } from "./components/Content/index.js";
 import React from "react";
 import WishList from "./components/WishList/index.js";
-import { ContactUsContent as Content } from "./components/Content/index.js";
 export default function ContactUs() {
   return (
     <>
