@@ -1,31 +1,54 @@
-import React from "react";
+import { Button, Typography } from "@mui/material";
 import {
-  FooterLink,
   ContentWrap,
-  LinksListItem,
-  LinksList,
-  Copyright,
   CopyrightListItem,
+  FooterLink,
+  LinksList,
+  LinksListItem,
 } from "./style.js";
+import { Auth } from "aws-amplify";
+import React from "react";
 
-export default function FooterLinks() {
+export default function FooterLinks({ isAdmin }) {
   return (
     <ContentWrap>
       <LinksList>
         <LinksListItem>
           <FooterLink href="/">Home</FooterLink>
         </LinksListItem>
-        {/* <LinksListItem>
+        <LinksListItem>
           <FooterLink href="/about">About</FooterLink>
-        </LinksListItem> */}
+        </LinksListItem>
         <LinksListItem>
           <FooterLink href="/contact">Contact Us</FooterLink>
         </LinksListItem>
         <LinksListItem>
           <FooterLink href="/watches">Pre-Owned Watches</FooterLink>
         </LinksListItem>
+        <LinksListItem>
+          {!isAdmin && (
+            <Button
+              variant="contained"
+              color="info"
+              sx={{ marginTop: "0.5em" }}
+              onClick={() => Auth.federatedSignIn()}
+            >
+              Login
+            </Button>
+          )}
+          {isAdmin && (
+            <Button
+              variant="contained"
+              color="info"
+              sx={{ marginTop: "0.5em" }}
+              onClick={() => Auth.signOut()}
+            >
+              Logout
+            </Button>
+          )}
+        </LinksListItem>
         <CopyrightListItem>
-          <Copyright>© 2021 Southwest Watches</Copyright>
+          <Typography align="center">© 2022 Southwest Watches</Typography>
         </CopyrightListItem>
       </LinksList>
     </ContentWrap>
