@@ -1,11 +1,11 @@
-import React from "react";
-import { FooterBox, ContentWrap, FooterTitle } from "./style.js";
-// import { Nav, NavDropdown, Navbar } from "react-bootstrap";
+import { ContentWrap, FooterBox, FooterTitle } from "./style.js";
 import FooterContactInfo from "./components/FooterContactInfo/index";
 import FooterLinks from "./components/FooterLinks/index";
 import FooterLogo from "./components/FooterLogo/index.js";
+import React from "react";
 import { useMediaQuery } from "@mui/material";
-export default function Footer() {
+
+export default function Footer({ isAdmin }) {
   const mediaQuery = useMediaQuery("(min-width:1100px)");
 
   return (
@@ -17,12 +17,12 @@ export default function Footer() {
         <FooterLogo />
         {mediaQuery ? (
           <>
-            <FooterLinks />
+            <FooterLinks isAdmin={isAdmin} />
             <FooterContactInfo />
           </>
         ) : (
           <>
-            <FooterContactInfo />
+            <FooterContactInfo isAdmin={isAdmin} />
             <FooterLinks />
           </>
         )}
